+++ DebrisDefragmentation/DebrisDefragmentation/Player.h
@@ -52,8 +52,8 @@ public:
 // 	void SetVelocity( DDVECTOR3 val ) { m_RigidBody.m_Velocity = val; }	
 // 	void SetAcceleration( DDVECTOR3 val ) { m_RigidBody.m_Acceleration = val; }
 	//DDVECTOR3 GetHeadDirection() { return m_HeadDirection; }
-	int GetGas() { return m_ClassComponent->GetFuel(); }
-	int GetOxygen() { return m_ClassComponent->GetOxygen(); }
+	float GetGas() { return m_ClassComponent->GetFuel(); }
+	float GetOxygen() { return m_ClassComponent->GetOxygen(); }
 
 private:
 	virtual void RenderItSelf();
